+++ src/main/java/org/reduct/client/ReductEntryClient.java
@@ -3,6 +3,7 @@ package org.reduct.client;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NonNull;
+import org.apache.commons.lang3.SerializationUtils;
 import org.reduct.client.config.ServerProperties;
 import org.reduct.common.EntryURL;
 import org.reduct.common.exception.ReductException;
@@ -12,6 +13,7 @@ import org.reduct.model.entry.Entry;
 import org.reduct.utils.Strings;
 import org.reduct.utils.http.HttpHeaders;
 
+import java.io.Serializable;
 import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
@@ -24,10 +26,10 @@ import static org.reduct.utils.http.Query.TIME_STAMP;
 @Getter(value = AccessLevel.PACKAGE)
 public class ReductEntryClient extends ReductClient implements EntryClient {
 
-
     protected final ServerProperties serverProperties;
     protected final HttpClient httpClient;
     protected final String token;
+
     ReductEntryClient(ServerProperties serverProperties, HttpClient httpClient, String accessToken) {
         this.httpClient = httpClient;
         this.serverProperties = serverProperties;
@@ -39,26 +41,51 @@ public class ReductEntryClient extends ReductClient implements EntryClient {
     }
 
     @Override
-    public String writeRecord(@NonNull Bucket bucket, @NonNull Entry<?> body) throws ReductException, ReductSDKException, IllegalArgumentException {
+    public String writeRecord(@NonNull Bucket bucket, @NonNull Entry<?> entry) throws ReductException, ReductSDKException, IllegalArgumentException {
         //TODO validation block
         if(Strings.isBlank(bucket.getName())
-           || Strings.isBlank(body.getName())
-           || Objects.isNull(body.getTimestamp())
-           || body.getTimestamp() <= 0
-           || Objects.isNull(body.getBody()))
+           || Strings.isBlank(entry.getName())
+           || Objects.isNull(entry.getTimestamp())
+           || entry.getTimestamp() <= 0
+           || Objects.isNull(entry.getBody()))
         {
             throw new ReductSDKException("Validation error");
         }
-        String timeStampQuery = String.format("?" + TIME_STAMP.getValue() +"=%d", body.getTimestamp());
-        URI uri = URI.create(getServerProperties().getBaseUrl() + String.format(EntryURL.WRITE_ENTRY.getUrl(), bucket.getName(), body.getName()) + timeStampQuery);
+        String timeStampQuery = getTimestampQuery(entry.getTimestamp());
+        URI uri = URI.create(getServerProperties().getBaseUrl() + String.format(EntryURL.WRITE_ENTRY.getUrl(), bucket.getName(), entry.getName()) + timeStampQuery);
         HttpRequest.Builder builder = HttpRequest.newBuilder()
                 .uri(uri)
-                .header(HttpHeaders.CONTENT_TYPE.getValue(), body.getBodyClass())
-                .POST(HttpRequest.BodyPublishers.ofByteArray(body.getByteBodyArray()));
+                .header(HttpHeaders.CONTENT_TYPE.getValue(), entry.getBodyClass())
+                .POST(HttpRequest.BodyPublishers.ofByteArray(entry.getByteBodyArray()));
 
         if(isNotBlank(getToken())) {
             builder.headers("Authorization", "Bearer %s".formatted(getToken()));
         }
-        return send(builder, HttpResponse.BodyHandlers.ofString());
+        return send(builder, HttpResponse.BodyHandlers.ofString()).body();
+    }
+
+    @Override
+    public <T extends Serializable> Entry<T> getRecord(Bucket bucket, Entry<?> entry) throws ReductException, ReductSDKException, IllegalArgumentException {
+        if(Strings.isBlank(bucket.getName()) || Strings.isBlank(entry.getName()))
+        {
+            throw new ReductSDKException("Validation error");
+        }
+        Long timestamp = entry.getTimestamp();
+        String timeStampQuery = Objects.isNull(timestamp) ? "" : getTimestampQuery(timestamp);
+        URI uri = URI.create(getServerProperties().getBaseUrl() + String.format(EntryURL.GET_ENTRY.getUrl(), bucket.getName(), entry.getName()) + timeStampQuery);
+        HttpRequest.Builder builder = HttpRequest.newBuilder()
+                .uri(uri)
+                .GET();
+        HttpResponse<byte[]> httpResponse = send(builder, HttpResponse.BodyHandlers.ofByteArray());
+        Object b = SerializationUtils.deserialize(httpResponse.body());
+        return (Entry<T>) Entry.builder()
+                .body((T) SerializationUtils.deserialize(httpResponse.body()))
+                .name(entry.getName())
+                .timestamp(httpResponse.headers().firstValue("x-reduct-time").map(Long::getLong).orElse(null))
+                .build();
+    }
+
+    private String getTimestampQuery(Long timestamp) {
+        return String.format("?" + TIME_STAMP.getValue() +"=%d", timestamp);
     }
 }