+++ src/main/java/org/reduct/model/entry/Entry.java
@@ -17,11 +17,6 @@ public class Entry<BODY_CLASS extends Serializable> {
     private Long timestamp;
     private final BODY_CLASS body;
 
-    public Entry(BODY_CLASS body, Long timestamp) {
-        this.timestamp = timestamp;
-        this.body = body;
-    }
-
     public String getBodyClass() {
         return body.getClass().getSimpleName();
     }